requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2
python-dotenv==1.0.0
python-dateutil==2.8.2

# Web technology detection
builtwith==1.3.4

# Optional but recommended dependencies
shodan==1.31.0
censys==2.2.9
dnstwist==20230918

# OSINT tools
sublist3r
theHarvester

# Additional utilities (if needed for theHarvester)
beautifulsoup4==4.12.2
lxml==4.9.3
netaddr==0.9.0
pyyaml==6.0.1

# PDF generation
reportlab==3.6.12
